requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-socketio
flask-migrate
flask-cors
python-dotenv
psycopg2-binary
cryptography
pydantic-settings
pydantic
flasgger
eventlet
